+++ Sources/SourceKitD/SourceKitD.swift
@@ -18,104 +18,49 @@ import TSCBasic
 import Dispatch
 import Foundation
 
-/// Wrapper for sourcekitd, taking care of initialization, shutdown, and notification handler
+/// Access to sourcekitd API, taking care of initialization, shutdown, and notification handler
 /// multiplexing.
 ///
-/// Users of this class should not call the api functions `initialize`, `shutdown`, or
+/// *Users* of this protocol should not call the api functions `initialize`, `shutdown`, or
 /// `set_notification_handler`, which are global state managed internally by this class.
-public final class SourceKitD {
-
-  /// The path to the sourcekitd dylib.
-  public let path: AbsolutePath
-
-  /// The handle to the dylib.
-  let dylib: DLHandle
-
+///
+/// *Implementors* are expected to handle initialization and shutdown, e.g. during `init` and
+/// `deinit` or by wrapping an existing sourcekitd session that outlives this object.
+public protocol SourceKitD: AnyObject {
   /// The sourcekitd API functions.
-  public let api: sourcekitd_functions_t
+  var api: sourcekitd_functions_t { get }
 
   /// Convenience for accessing known keys.
-  public let keys: sourcekitd_keys
+  var keys: sourcekitd_keys { get }
 
   /// Convenience for accessing known keys.
-  public let requests: sourcekitd_requests
+  var requests: sourcekitd_requests { get }
 
   /// Convenience for accessing known keys.
-  public let values: sourcekitd_values
+  var values: sourcekitd_values { get }
 
-  /// Lock protecting private state.
-  let lock: NSLock = NSLock()
-
-  /// List of notification handlers that will be called for each notification.
-  private var _notificationHandlers: [WeakSKDNotificationHandler] = []
-
-  var notificationHandlers: [SKDNotificationHandler] {
-    lock.lock(); defer { lock.unlock() }
-    return _notificationHandlers.compactMap { $0.value }
-  }
-
-  public enum Error: Swift.Error {
-    /// The service has crashed.
-    case connectionInterrupted
-
-    /// The request was unknown or had an invalid or missing parameter.
-    case requestInvalid(String)
-
-    /// The request failed.
-    case requestFailed(String)
-
-    /// The request was cancelled.
-    case requestCancelled
-
-    /// Loading a required symbol from the sourcekitd library failed.
-    case missingRequiredSymbol(String)
-  }
-
-  public init(dylib path: AbsolutePath) throws {
-    self.path = path
-    #if os(Windows)
-    self.dylib = try dlopen(path.pathString, mode: [])
-    #else
-    self.dylib = try dlopen(path.pathString, mode: [.lazy, .local, .first])
-    #endif
-    self.api = try sourcekitd_functions_t(self.dylib)
-    self.keys = sourcekitd_keys(api: self.api)
-    self.requests = sourcekitd_requests(api: self.api)
-    self.values = sourcekitd_values(api: self.api)
-
-    self.api.initialize()
-    self.api.set_notification_handler { [weak self] rawResponse in
-      guard let self = self else { return }
-      self.lock.lock()
-      let handlers = self._notificationHandlers.compactMap(\.value)
-      self.lock.unlock()
-
-      let response = SKDResponse(rawResponse, sourcekitd: self)
-      for handler in handlers {
-        handler.notification(response)
-      }
-    }
-  }
-
-  deinit {
-    self.api.set_notification_handler(nil)
-    self.api.shutdown()
-    // FIXME: is it safe to dlclose() sourcekitd? If so, do that here. For now, let the handle leak.
-    dylib.leak()
-  }
-
-  /// Adds a new notification handler (referenced weakly).
-  public func addNotificationHandler(_ handler: SKDNotificationHandler) {
-    lock.lock(); defer { lock.unlock() }
-    _notificationHandlers.removeAll(where: { $0.value == nil })
-    _notificationHandlers.append(.init(handler))
-  }
+  /// Adds a new notification handler, which will be weakly referenced.
+  func addNotificationHandler(_ handler: SKDNotificationHandler)
 
   /// Removes a previously registered notification handler.
-  public func removeNotificationHandler(_ handler: SKDNotificationHandler) {
-    lock.lock(); defer { lock.unlock() }
-    _notificationHandlers.removeAll(where: { $0.value == nil || $0.value === handler})
-  }
+  func removeNotificationHandler(_ handler: SKDNotificationHandler)
+}
+
+public enum SKDError:Error {
+  /// The service has crashed.
+  case connectionInterrupted
+
+  /// The request was unknown or had an invalid or missing parameter.
+  case requestInvalid(String)
+
+  /// The request failed.
+  case requestFailed(String)
+
+  /// The request was cancelled.
+  case requestCancelled
+
+  /// Loading a required symbol from the sourcekitd library failed.
+  case missingRequiredSymbol(String)
 }
 
 extension SourceKitD {
@@ -142,7 +87,7 @@ extension SourceKitD {
   public func send(
     _ req: SKDRequestDictionary,
     _ queue: DispatchQueue,
-    reply: @escaping (Result<SKDResponseDictionary, Error>) -> Void
+    reply: @escaping (Result<SKDResponseDictionary, SKDError>) -> Void
   ) -> sourcekitd_request_handle_t? {
     logAsync { _ in req.description }
 
@@ -172,13 +117,7 @@ extension SourceKitD {
   }
 }
 
-public protocol SKDNotificationHandler: class {
+/// A sourcekitd notification handler in a class to allow it to be uniquely referenced.
+public protocol SKDNotificationHandler: AnyObject {
   func notification(_: SKDResponse) -> Void
-}
-
-struct WeakSKDNotificationHandler {
-  weak private(set) var value: SKDNotificationHandler?
-  init(_ value: SKDNotificationHandler) {
-    self.value = value
-  }
 }
